extracting some sub-interfaces
