+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v1/ConfigurationService.scala
@@ -1,0 +1,13 @@
+// Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.participant.state.index
+
+import com.daml.ledger.participant.state.v1.Configuration
+
+import scala.concurrent.Future
+
+trait ConfigurationService {
+  def getLedgerConfiguration(): Future[Configuration]
+
+}
