+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v1/IdentityService.scala
@@ -1,0 +1,12 @@
+// Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.participant.state.index
+
+import com.daml.ledger.participant.state.v1.LedgerId
+
+import scala.concurrent.Future
+
+trait IdentityService {
+  def getLedgerId(): Future[LedgerId]
+}
