+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v1/IndexService.scala
@@ -5,29 +5,32 @@ package com.daml.ledger.participant.state.index.v1
 
 import akka.NotUsed
 import akka.stream.scaladsl.Source
+import com.daml.ledger.participant.state.index.{
+  ConfigurationService,
+  IdentityService,
+  PackagesService,
+  TransactionsService
+}
 import com.daml.ledger.participant.state.v1._
-import com.digitalasset.daml.lf.data.Ref.PackageId
 import com.digitalasset.daml.lf.data.Time
-import com.digitalasset.daml.lf.transaction.BlindingInfo
 import com.digitalasset.daml.lf.value.Value
 import com.digitalasset.daml.lf.value.Value.{AbsoluteContractId, ContractInst}
-import com.digitalasset.daml_lf.DamlLf.Archive
 import com.digitalasset.ledger.api.domain.TransactionFilter
 
 import scala.concurrent.Future
 
-trait IndexService {
-  def listPackages(): Future[List[PackageId]]
-  def isPackageRegistered(packageId: PackageId): Future[Boolean]
-  def getPackage(packageId: PackageId): Future[Option[Archive]]
+trait IndexService
+    extends PackagesService
+    with ConfigurationService
+    with TransactionsService
+    with IdentityService {
 
-  def getLedgerConfiguration(): Future[Configuration]
-
-  def getLedgerId(): Future[LedgerId]
-
+  //TODO: review these where are they used and what concerns they fulfill
   def getLedgerBeginning(): Future[Offset]
+
   def getLedgerEnd(): Future[Offset]
 
+  //TODO gabor: why is this seperate from the transactions, isn't it a problem that they can be out of sync?
   def getLedgerRecordTimeStream(): Source[Time.Timestamp, NotUsed]
 
   def lookupActiveContract(
@@ -39,28 +42,12 @@ trait IndexService {
       filter: TransactionFilter
   ): Future[ActiveContractSetSnapshot]
 
+  //TODO gabor: why the update, when is it used?
   def getActiveContractSetUpdates(
       beginAfter: Option[Offset],
       endAt: Option[Offset],
       filter: TransactionFilter
   ): Source[AcsUpdate, NotUsed]
-
-  // FIXME(JM): Cleaner name/types for this
-  // FIXME(JM): Fold BlindingInfo into TransactionAccepted, or introduce
-  // new type in IndexService?
-  def getAcceptedTransactions(
-      beginAfter: Option[Offset],
-      endAt: Option[Offset],
-      filter: TransactionFilter
-  ): Source[(Offset, (TransactionAccepted, BlindingInfo)), NotUsed]
-
-  /*
-  def getTransactionById(
-      ,
-      transactionId: TransactionId,
-      requestingParties: Set[Party]
-  ): Future[Option[TransactionAccepted]]
-   */
 
   def getCompletions(
       beginAfter: Option[Offset],
