+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v1/PackagesService.scala
@@ -1,0 +1,15 @@
+// Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.ledger.participant.state.index
+
+import com.digitalasset.daml.lf.data.Ref.PackageId
+import com.digitalasset.daml_lf.DamlLf.Archive
+
+import scala.concurrent.Future
+
+trait PackagesService {
+  def listPackages(): Future[List[PackageId]]
+  def isPackageRegistered(packageId: PackageId): Future[Boolean]
+  def getPackage(packageId: PackageId): Future[Option[Archive]]
+}
